Finish pulling out passes & utility functions from document converter
