+++ src/passes/remove-unnecessary-waits.ts
@@ -14,17 +14,56 @@
 
 import * as estree from 'estree';
 
-import {getMemberName, toplevelStatements} from '../util';
+import {getMemberName, getTopLevelStatements} from '../document-util';
+
+/**
+ * Get a reference to the callback in a
+ * `document.addEventListener("WebComponentsReady", ...)` call expression.
+ */
+function getWebComponentsReadyCallback(
+    memberName: string|undefined,
+    callExpression: estree.CallExpression): estree.Node|undefined {
+  const args = callExpression.arguments;
+  if (args.length !== 2) {
+    return;
+  }
+  if (!(memberName === 'addEventListener' ||
+        memberName === 'document.addEventListener')) {
+    return;
+  }
+  const [eventName, callback] = args;
+  if (eventName.type !== 'Literal' ||
+      eventName.value !== 'WebComponentsReady') {
+    return;
+  }
+  return callback;
+}
+
+/**
+ * Get a reference to the callback in a
+ * `document.addEventListener("HTMLImports.whenReady", ...)` call expression.
+ */
+function getHtmlImportsReadyCallback(
+    memberName: string|undefined,
+    callExpression: estree.CallExpression): estree.Node|undefined {
+  const args = callExpression.arguments;
+  if (args.length !== 1) {
+    return;
+  }
+  if (memberName !== 'HTMLImports.whenReady') {
+    return;
+  }
+  return args[0];
+}
+
 
 /**
  * Unwrap the bodies of any listeners for the `WebComponentsReady` event, or
  * callbacks passed to `HTMLImports.whenReady` as the closest equivalent in
  * modules world happens by default.
- *
- * @param program The program. Is mutated.
  */
 export function removeUnnecessaryEventListeners(program: estree.Program) {
-  for (const nodePath of toplevelStatements(program)) {
+  for (const nodePath of getTopLevelStatements(program)) {
     const statement = nodePath.node;
     if (statement.type !== 'ExpressionStatement' ||
         statement.expression.type !== 'CallExpression') {
@@ -53,36 +92,4 @@ export function removeUnnecessaryEventListeners(program: estree.Program) {
     }
     nodePath.prune();
   }
-}
-
-function getWebComponentsReadyCallback(
-    memberName: string|undefined,
-    callExpression: estree.CallExpression): estree.Node|undefined {
-  const args = callExpression.arguments;
-  if (args.length !== 2) {
-    return;
-  }
-  if (!(memberName === 'addEventListener' ||
-        memberName === 'document.addEventListener')) {
-    return;
-  }
-  const [eventName, callback] = args;
-  if (eventName.type !== 'Literal' ||
-      eventName.value !== 'WebComponentsReady') {
-    return;
-  }
-  return callback;
-}
-
-function getHtmlImportsReadyCallback(
-    memberName: string|undefined,
-    callExpression: estree.CallExpression): estree.Node|undefined {
-  const args = callExpression.arguments;
-  if (args.length !== 1) {
-    return;
-  }
-  if (memberName !== 'HTMLImports.whenReady') {
-    return;
-  }
-  return args[0];
 }
